requirements.txt

Flask==3.0.3
joblib==1.4.2
tldextract==5.1.2
python-whois==0.9.4
requests==2.32.3
